Unused code from emotion detector demo
For my Lunch & Learn talk at OctoML, I first attempted to build an Arduino demo that
would use the camera to identify the facial expression of a person, and then display
an emoji to match. This was complicated, as I ended up having to use a floating point
model as well as soldered-on QSPI flash with a Teensy 4.1. However, even once I got
it working I ran into the problem that the model was just plain flaky. Thus, the idea
was scrapped. In case anyone wants any of that code, here it is. Use at your own risk.